SPP1-39 Hide flag updates within Scan.rfi
Previously after computing the flags with `Scan.rfi` the pipeline was
explicitly storing the computed flags in a shared memory flag array with
`store_flags`.
As storing the flags is always desirable after computing them it
makes more sense to combine these actions into a single step under
`Scan.rfi`
