+++ katsdpcal/reduction.py
@@ -62,71 +62,6 @@ def init_flagger(parameters, dump_period):
     return calib_flagger, targ_flagger
 
 
-def slot_slices_to_time_slices(slices):
-    """Convert a list of slot slices to time slices"""
-    start = 0
-    for s in slices:
-        step = s.stop - s.start
-        yield slice(start, start+step)
-        start = start+step
-
-
-def slots_slices(slots):
-    """Compresses a list of slot positions to a list of ranges (given as slices).
-
-    This is a generator that yields the slices
-
-    Example
-    -------
-    >>> list(_slots_slices([2, 3, 4, 6, 7, 8, 0, 1]))
-    [slice(2, 5, None), slice(6, 9, None), slice(0, 2, None)]
-    """
-    start = None
-    end = None
-    for slot in slots:
-        if end is not None and slot != end:
-            yield slice(start, end)
-            start = end = None
-        if start is None:
-            start = slot
-        end = slot + 1
-    if end is not None:
-        yield slice(start, end)
-
-
-def store_flags(slots, scan_slice, flag_array, scan, auto_ant=False):
-    """Store scan flags in the shared memory flag buffer
-
-    Inputs
-    ------
-    slots: list of slots
-        buffer slots currently being processed by the pipeline
-    scan_slice: slice
-        slice of the current scan
-    flag_array: :class:`np.ndarray` of uint8, shape (ntimes, nchans, npols, nbls)
-        shared memory array containing flag data
-    scan: :class:`Scan`
-        scan corresponding to the scan_slice
-    cross: bool
-        cross-correlations or auto-correlations
-    """
-    # ******* Fancy indexing will not work here, as it returns a copy, not a view *******
-    # all selections on the shared flag array must be done using slices
-    if auto_ant:
-        mask = scan.ac_mask
-        corr = scan.auto_ant
-    else:
-        mask = scan.xc_mask
-        corr = scan.cross_ant
-
-    mask_slice = slice(np.where(mask)[0][0], np.where(mask)[0][-1]+1)
-    slot_slices = list(slots_slices(slots[scan_slice]))
-    time_slices = list(slot_slices_to_time_slices(slot_slices))
-    for slots, times in zip(slot_slices, time_slices):
-        flag_array[slots, :, :2, mask_slice] = corr.tf.auto_pol.flags[times].compute()
-        flag_array[slots, :, 2:, mask_slice] = corr.tf.cross_pol.flags[times].compute()
-
-
 def get_tracks(data, telstate, dump_period):
     """Determine the start and end indices of each track segment in data buffer.
 
@@ -661,7 +596,7 @@ def pipeline(data, ts, parameters, solution_stores, stream_name, flag_array, slo
                  parameters['bls_lookup'], target,
                  chans=parameters['channel_freqs'],
                  ants=parameters['antennas'],
-                 refant=refant_ind,
+                 flag_array=flag_array, slots=slots, refant=refant_ind,
                  array_position=parameters['array_position'], logger=logger)
         if s.xc_mask.size == 0:
             logger.info('No XC data - no processing performed.')
@@ -687,7 +622,6 @@ def pipeline(data, ts, parameters, solution_stores, stream_name, flag_array, slo
         if any(k.endswith('cal') for k in taglist):
             logger.info('Calibrator flagging')
             s.rfi(calib_flagger, sensors=sensors)
-            store_flags(slots, scan_slice, flag_array, s)
             # Set a reference antenna for this cbid if one isn't already set
             if s.refant is None:
                 set_refant(s, ts, parameters, sensors)
@@ -700,7 +634,6 @@ def pipeline(data, ts, parameters, solution_stores, stream_name, flag_array, slo
         if any('bfcal' in k for k in taglist):
             logger.info('Calibrator flagging, auto-correlations')
             s.rfi(calib_flagger, auto_ant=True, sensors=sensors)
-            store_flags(slots, scan_slice, flag_array, s, auto_ant=True)
             # ---------------------------------------
             # K solution
             logger.info('Solving for K on beamformer calibrator %s', target_name)
@@ -979,7 +912,6 @@ def pipeline(data, ts, parameters, solution_stores, stream_name, flag_array, slo
                 # flag calibrated target
                 logger.info('Flagging calibrated target {0}'.format(target_name,))
                 s.rfi(targ_flagger, sensors=sensors)
-                store_flags(slots, scan_slice, flag_array, s)
 
             # summarize corrected data for data with cal tags
             logger.info('Averaging corrected data for %s:', target_name)
